+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
 * Unreleased
+* 0.6 (2021-02-19)
+    * Add `libraries/EpoxyFS`, an implementation of the ESP8266 `FS` file system
+      that runs on Linux and MacOS.
+    * Add `delayMicroSeconds()`, `WCharacter.h`, and stub implementations of
+      `IPAddress.h`, `SPI.h`, by Erik Tideman (@ramboerik), see
+      [PR #18](https://github.com/bxparks/EpoxyDuino/pull/18).
+    * Add `libraries/EpoxyPromEsp` (EEPROM using the API of ESP8266/ESP32).
+    * Add `libraries/EpoxyPromAvr` (EEPROM using the API of AVR processors).
 * 0.5 (2021-01-21)
     * **Breaking Change** Change project name to "EpoxyDuino". The
       `UNIX_HOST_DUINO` macro is replaced with `EPOXY_DUINO`, and the